Refactoring the get_dataset and scaling part with the intention to reduce comlexity
